([*] fix the way we libricate (.. 2574))
For those sidesystems that use both a `LIB_`
instance and the `Foo._lib` reader, we were doing
it wrong before. in the old way, accessing the
`_lib` reader would trigger the autoloading of
the library file. ** while this file was loading **,
it would again produce another library shell and
assign it to the `LIB_` constant. this process itself
would assign to the `@lib` ivar. then when this file
was finished loading, the `@lib` ivar would again get
assigned to the other lib shell that was built
unnecessarily.

now we build only one lib shell and store it in a
protected constant.

(2 in 9.82)