([hl]<=[st] 2/2 re-unify filesystem "tree walk" (.. 2596))
Move [br]’s newer, simplified “filesystem walk” into
place inline with the old one, which is effectively
commented-out for now.

the old one had been offlined as long as doc test
had been (which had become its only client) so we
are taking the opportunity to re-unify this node.

the real “step 3” of this series will be in the next
cluster of commits, where we rewrite doc-test’s
recursive action to use our new filesystem “tree walker”
and then prune the older code by this.

(2 in 12.25)