([ts]           regenerate 1/5 (3/4) (renames/removes)..)
we got rid of code fixtures that were within the prodution code
because we were sick of looking at it. the newer implementation
has less architecture-specific requirements so we don't need to
mock stuff like this.

move action from the legacy "regret" taxonomic application node
and into "doc-test". the "intermediates" utility is not exactly
part of what "doc-test" does, but it feels right to group these
as siblings, because they both generate test-related files, and
are probably the only ones that ever will.