([tr] interlude 3/3: kill it with fire)
Synopsis: FINALLY getting rid of some of this
unmanageable ball of tangle. we were having
the weirdest attachment anxiety and guilt about
doing this, but those feelings lasted for
about 10 minutes.

all the experimentation became a barrier to
business. no time to wait for the “official”
assessment. we know we were doing it wrong
and we need to fix it now so we can ping
again.