([sg] 12.D spike first full criteria parse tree (.. 2784))
Synopsis: wow so although there is a lot of mess here yet, we
  begin to see the potential fun with this mechanic.
  the main thing here is the entrance of a domain adapter and
  the crazy actor for parsing entire criteria..

Ramblings: halfway through this step, and after having made the
  overwrought “item from matrix” function over in [mh], we started
  thinking, “isn’t it ugly that we hard-coded the verb to be
  only one word long?” ..

  one we had it that the verb is inflected for 3rd person singular
  in the query (“Foos that have ..”) but the “lemma” form is not
  that, we decided to employ our new `interpret_out_of_under`
  pattern, where as the expression agent we pass our (born in
  this step) “grammatical context” object, which in this case
  holds only the binary info of whether the verb must agree with
  a plural or singular number for the subject.

  although it remains kind of thick and somewhat ancient, the
  [hl] EN::POS library came in handy for this, to conjugate the
  verbs to agree with the grammatical context so we can correctly
  parse the input based on what number (hypothetically) the
  subject noun has (currently it requires a plural number).

  then this got our head spinning a tiny bit: contrast:

    ‘users who are named ”Johnson”’  -vs-
    ‘a user who is named “John”’  -vs-
    ‘the user who is named “John”’

  the first production is close to the model case for this step
  and probably the near future work in this space. however,
  the second and third productions have a semantic difference
  which could be interpreted by the system: the use of the
  indefinite article (“a” or “an’) could be used to express
  the desire for *any* first item matching the criteria. the
  last example, however, makes the **expressed assumption**
  that there is only one such item matching this query. this
  could prove useful in the way future for the top-secret project
  codenamed “sudocode”..

Other ramblings: futhermore we began to see a redundancy with what
  we are calling [#005] for lack of any name yet (something like
  “plastic clause boundaries”).. our most recent implementation of
  a parser for:

    “A”  -or-
    “A or B”  -or-
    “A or B or C”  -but not-
    “A or B and C”

  we now have something like this implemented something like 3 ways
  in the code. we are close but not quite there with knowing what the
  “elegant” general solution (if any) to this is. this commit, then,
  is just a sanity commit to lock down what we have so far, so we can
  test against it all when we try to tighten this all up.

  [#013] marks one area we would like to clean up

“Tesselating phrase boundaries”?

imagine that our hacky “verb phrase head” structure is:

    “is tagged with ..”

the head verb there is “is” which becomes “are” in:

    “things that are tagged with ..”

etc, that we would use in parsing queries etc. that much is kind of
“straightforward” and is something we have a prototype of here in this
commit. we bring this example up to a) show how it might be useful
to support multiple (er) word strings in the expression of this
“verb phrase head”, as opposed to just supporting single verbs like
“is”, “has”, “costs” etc.

but (b) even more cool than that is that this three-word phrase-head
lets us try to find our general solution to enabling the “tesselating
phrase boundaries” idea. parse these sentences with your mind:

    ‘things that are tagged with “foo” or “bar”’
    ‘things that are tagged with “foo” or with “bar”’
  ! ‘things that are tagged with “foo” or “bar” and “baz”’

by our rules, there is no difference in the first two (and under
the “no synonyms” rubric, we would like to avoid supporting both
of them but we will anyway). but more importantly, in the third
example, by our rules this is unrecoverably ambiguous. in contrast
to all programming languages we have this much knowledge of, in these
cases we refuse to make an arbitrary “judgment of solomon” decision
(bad reference) in picking one over the other.  i.e “it doesn’t make
sense” to pick AND over OR (or vice versa) in resolving this precedence
debate.

and yes programming languages (that we yet know about) resolve these
issues partially with precedence rules and partially with parenthesis.
WELL: as a sort of OCD gimmick, we are going to say that we don’t want
to rely on parenthesis because spoken language doesn’t rely on
parenthesis and seems to work out with avoiding ambiguity for most
day-to-day pragmatics of communication.

SO, in contrast to that last example, try reading this instead:

    ‘things that are tagged with “foo” or “bar” and with “baz”’

it’s subtle (and maybe so subtle that it’s useless), but adding that
extra “with” in there saves the day for us, in our minds.

    ‘nodes that are tagged with “urgent” or “important” and with
       “project foo” or tagged with “crisis”’

ok that example is super bad, but the idea is that we can have up
to 3 (or so?) levels of depth of nested nodes in our boolean tree,
because of the varying number of context words we can re-introduce
to break out of other contexts.  anyway, it’s just a thought.

yeah we are sick of it already..

(2 in 16.24)