([br]  2/N    * property stack renames..)
Justification for the renames:

It's time we promote this strange but enduring artifact out of
"entity/properties stack" and move it up upwards into its own top-ish
library. but for semantics and taxonomics we went ahead and started
a toplevel "property" node; with justification provided by the heavy
use it gets as a library for building common property-related
events. it was smell to be using the "properties stack" node everywhere
for accesssing these event builders.
_