([br]  1/N  • begin massive simplification (one remove)..)
Ramblings: the overall goals of this large interlude will be expressed
 in the upcoming commits that it is composed of.

this large interlude is composed of a series of "clusters". every
"cluster" is composed of a series of a series of one or more
nodes where a node is a commit or another cluster.

in this cluster we try something new: the indentation of the leading
dots (and what "glyph" we use) indicate the depth of that node
in the cluster tree.

this commit specifically:

([br] interlude  1/N: remove orphan test support)
_