([br]  3/N    * property stack rebuild.. (( 43 )) )
Synopsis: this will necessarily be a very fragmented series of steps,
but here is a start. "property stack" is simplified somewhat by the
simplification of "entity". in other ways it has to implement its
own methods where before it didn't.

(this cluster happened because we wanted to have a clean tree to
test "entity" against, and "pstack" was in that tree.)

introduction to the broader cluster will be next.
_