([ts] 13/N    * integration ..)
an argument arity of “custom” now means that you will define a
`foo_bar=`-style method for your field to do its parsing from the
polymorphic upstream. the “permute” does not currently nor did in
the past implement the “permutations” field in this manner, so we
are not sure why we did it that way.

(rather the front client does not treat this field as isomorphic,
it must produce the “permutations” stream “manually”.)

implementing the stuff near “enum” here made us see our next big
thing ..

  • for IO spy just use methodic actor not entity.
_