([br] 4/N fill (.. 2659))
quite naturally it becomes a tree of abstractions ..

• broke out specific “max share meter” spec
  (not yet re-integrated)

• re-wrote fill (alongside the freshly re-written “lipstick”..)

This commit itself underwent several rewrites around its architecture.
we read up on Martin Fowler near a definition of Bad Design, and
pondered over and over again how we could avoid having the higher
level modules dependent on the lower level ones.

The main innovation we came up with here is that the field structures
themselves are implemented with a plugin-like architecture of
strategies, each of which can add parseable terms to the syntax,
which is how for example the `fill` term can by parsed by the
`field` term without the latter knowing about the former.

  • [pl] abstract out the “shared parse” logic (something we’ve done
         similarly elsewhere several times) so that both field and
         row receiver (or whichever) can etc.

(2 in 13.1, 13.06)