([sc] unit 05 step 2: sortable list second rewrite ..)
our big paradigm shift for this SECOND FULL REWRITE was that we only
ever change the Y transform of the transorm matrix of the elemeents.
we do not and should not want to deal with absolute positioning, as
that doesn't give us content flow.

(HAPPY?)
(all but animations)
(one more animation to do)
best yet ever. remaining issue: z index. and that's ALL
(changes for the demo)
_