([mt] 1/N: a "join point" - clear these back to empty)
.. it will be amazing if we pull this off: myterm is about to undergo
a full rewrite. we want it to join into our ecosystem from outside.
towards this end we create this commit as a "join point" for it to
go from outside to in..

.. this is for our one-day experiment with [sl]: we are going to
attempt to "re-acquisition" the old "myterm" as one that lives under
[tmx]: this will happen by having a commit in tmx and a commit in the
sovereign [mt] that both leave the project in an identical state (this
state in this commit). then, each commit that we make here in [tmx]
towards [mt], the sovereign [mt] will be a downstream recipient of
those changes. this way, [mt] can live under [tmx] for day-to-day
development, but as it is distributed it can live in its own dedicated
repository so that it can live life like a normal small gem in the
 git[hub] ecosystem.

the trick will be that a) we must make sure that the soverign [mt] ends
up with a commit that leaves it in this identical state (file for file
and byte for byte)..
_