([*]  better distinguish “qualified knownness” from etc (.. 2591))
Catalyst: it got annoying because some normalizers expected to have
a qualified knownness and others just thought they had a known known..
we now make this official that the normal normalizer takes a qualified
knownness. to reinforce this, we are refactoring out all names like
“arg” and “trio” as possible.

we again rename “is_known” to “is_known_known” - a qualified knownness
might be used to represent a known unknown in which case it still
represents a known but not a known known. the name "is_known" glosses
over this distinction hence we eliminated it.

(2 in 14.44, 14.8)