(         [se]       waypoint 2.. (.. 3127))
Synopsis: spike a bunch of CLI-related work from mid-january. although
the bulk of the branch-commit ("stash"-like) has been changed beyond
recognition from the original spike, we have retained the original
commit date from this stash both for posterity and to challenge
visualizations like [gv] with a real-world use-case.

midway through re-integrating this work into the milestone sequence,
we enountered and then (theoretically) solved the problem that is
explored in [#ze-027] (what at the moment we refer to as the
"socialism"/"individualism" dichotomy).

to realize [#ze-027] will require some "revisionist" steps that drop
us back into phase 1. after deliberation and investigation, we decided
that rather that postpone this spike any futher, it is better to get
it all out now and re-integrate the new (subject) theory in this
subsequent pass that is up next ..

(2 in 16.65)