(step 6/6 [ze]<-[*]  waypoint 1: beginnings of availability.. (.. 2970))
Process:

We now introduce (to ourselves) a new project-management-y structural
unit: the "waypoint":

waypoints are like mini-milestones. (we considered names like "goal",
"objective", "deliverable", etc.)

we are now defining "waypoints" as special steps that we can predict
beforehand with some (but not certain) certainty. (whereas ordinary
"steps" are one at a time and emergent).

we will define (and number) waypoints within the context of each
milestone. so the known waypoints for this milestone are:

  1 - something like "availability" for operations
  2 - availability for non-operations
  3 - a mask/mapper that disables particular associations by name
      for the "intent" of API

  4 - non-proc operations..

.. and hopefully with those 4 done, we can close "milestone 1"

This commit (chrono order):

  • finally get rid of a nasty asymmetry -
    now you must be compound to have operations..

  • we once had support for something like transitive operations
    in the API (perhaps for shot-gunning getters and setters).
    now that these are built-in to the generated API's (see [#012])
    we have removed this implementation and API detail. we will revisit
    this anew if we ever wish we had them, rather than drag the old
    cruft along. (the specs now demonstrate the new way of using
    the out-of-the-box getter/setters.)

  • [ca] (etc)

(2 in 12.98)