(step 1/N [ac]<-[*]  no more hot eventmodel (for now).. (.. 2997))
Synopsis: while experimentally stripping away implementation for the
 “hot” event model [#ac-006], we also improved contextual expression
 of (expression) emissions in the JSON unserializing node.

Minor Side-effects:

  • [ba] oops in a merge we must have un-removed this one file which two
    months later was assimilated into the “small time actors” node

Narrative: we were trying to use a testing instrument from [ze] over
in [mt]. it occurred to us we didn’t even know what eventmodel we
wanted (hot or cold). as we followed along the code in the test support
that builds the root ACS, we found it confusing to have to switch back
and forth from hot to cold in our heads.

we broke for dinner, played some DDR, and then out of nowhere while
we were going to get more quarters, a voice popped into our head
about the issue. it said to us, “your model is not your application.”
that was it. the model is the model. its job is to hold data. (and why
should every node in a potentially large structure have to worry about
its own handle on an event handler?)

we need to go write this up in a doc..

..salient in this commit is the many occurrences of definitions for
`new_cold_root_ACS_for_expect_root_ACS`. having this method definition
repeated so many times is certainly a violation of DRY, so here’s the
justification:

  • the calling library sends this and not `new` to drive home the
    important point of [#002]Tenet1 (the ACS doesn’t ever call `new`).

  • we don’t make a proc called `Be_root_ACS` or the like to drive home
    the point (i.e remind ourselves of it) that there is not and should
    not be a large initial API (“setup cost”) to use the [ac] on your
    root class.

Primordial list of current known waypoints:

  waypoint 1 - show current adapter (not available)
  waypoint 2 - list adaters
  waypoint 3 - set adapter EESH

(2 in 13.02)