(         [my]<-[ac] waypoint 5: injected compound (.. 3022))
Synopsis: implement injection of a compound node.

In the added test cases it is demonstrated that syntax is added
dynamically that reveals new features added by the selected adapter.

Side-effects: Testament to the clean-up, the imagemagick adapter
node is now much more readable (and short for the moment), with the
advent of the new “adapter adapter” node.

  • let’s not be afraid of dedicated sub-nodes for components.
    (a collection controller for fonts)

  • get rid of old “get” and “set” component operations - out of box now

  • [ac] (because we’re never gonna turn this into 
    “Reader::Writer”, let’s make the name platform-normal.)

  • [ac] add support for defining description of op even
    though it won’t see coverage for a while..

  • [ac] add “detection” of association definition 
    experimentally (covered only by [my] for now)

(2 in 15.4)