(         [se]       reach Milestone 3.. (.. 3075))
Chatter: wow! smooth sailin'.. (this is the combined efforts of reaching
waypoints 3 *and* 4, both of which were pretty straighforward.)

right a wrong that we once did (the “replace” node is now an
ordinary session implementing an operation). we came up with a nasty
(fun) potential “improvement” to the API API that we’re gonna have to
save for later (dog-eared with [#020]).

(2 in 8.22)