(    sub-waypoint 2/N ([hu]<=[br])<-[*] more c15n (.. 3085))
Synopsis: tear out c15n-related logic from [br] CLI and shove it into
the new home for c15n. tried to make it pretty but some of it is at
odds w/ what we just set up. (redundancy of similar expressions
reached thru different mechanisms, the graph isn’t used as well
as it could be..) but the good things here are that:
  • [br] is using an API for the c15n and
  • what [br] used to do is now exposed

• [br] removed “redundancy filter” (not covered anywhere)

(2 in 16.87)