(interlude 2/K: wip everything.. (( 70 )) )
..the oldest still living attributes-related node in this universe
(as far as we care to know right now) will now become the dedicated
“default meta-attributes” node (which before this, for “attributes”
was in-like (i.e “stowaway”) with the “lib” node..)

in this commit we break the old parameters lib (used by [ca], [css],
[tm]) for the purpose of assimilating it progressively into our
hub stream.

we will likely eventually delete some of these tests, but meh for now.
_