(interlude 8/L: integrate [tm]->[fi] but read this.. (( 220, 116 )) )
decimate and anemicize “bounder”. there are known holes with this
restoration near reading the names of bound attributes, but since
they aren’t covered we’re not going to bring them back blindly, nor
are we going to add new coverage today. tombstones mark the old code
for when (if ever) we take this on.

• [fi] added `normalize_session` and restored old n11n test.
       broke up formerly “Parse” to drive also plain old n11ns w/o parse
_