(interlude 8/M: improve parameter handling in [ta].. (.. 3013))
..riffing off of the adding of support for optionals in tasks, we
discovered an architectural smell that wanted improving:

..it used to be we saw every parameter “requirement” as a different
dependency association to manage. now we have one pseudo-node that
tries to resolve all requisite parameters as one node. this way we
A) re-use the excellent shiny new [fi] facilities and B) get
aggregate and standard-issue messages about missing requireds
(instead of the graph resolution failing on the first missing
required parameter.)

we snuck in probable support for defaults too..

• [fi] build out normalization API to a more granular level..

(2 in 14.76)