m1 4/M: pre-reasonification (mostly renames).. (.. 2969)
Synopsis: generalize the “missing” event to handle unmet dependencies
and the more complex explanations we will have there. “reasons” will
be used later; this is just setting the stage.

Note: tons of renames of “property” to “attribute”. although we could
have stuck with the name “property” (we have the option of passing any
arbitrary lemma to the event builder); we chose to adopt the new
default for consistency, and out of deference to Chen et. al. and the
deeper dream..

  • rename `on_missing_required` to `on _reasons’
  • rename `value_reader_proc` to 	`evaluation_proc`
  • rename `knownness_for` to `evaluation_for`

  • [fi] don’t crowd the method namespace *at all* of meta-attributes

(2 in 16.87)