m4 2/J: spike help screens for compound frames.. (.. 3084)
Narrative: a long time coming, this is the final form of about a third
of a large-ish stash that we spiked, largely had working, and then
stashed away when we came up with the 9-point milestone sequence
MONTHS ago EGADS..

in the interceding time between its inception and now, we read a
smattering of pieces from
_Patterns, Principles and Practices of Domain-Driven Design_
[#sl-157] (whose influence first hit this code at this commit minus 5.)
from this we see a spike of “unified language” type documents being
created here.

this new approach has been a boon to our code near the ubiquitous
term “node” - there was vagueness and variability there making code
hard to decipher at a glance. as a first step of amelioration, we have
kept the archipelago of representation (because each island has its
own purpose), but we have unified them conceptually under this idea
of the “4-stage lifecycle of a node”. now we know the importance of
chasing the right name: it’s (typically) either a node ticket or a
formal node. it should never be just “node”.

(we have even normalized the variable names to use for these.)

  • wow we accidentally eliminated the unfortunate term “associationesque”

On adding help (all [ze]):

  • jostle the syntax parsing a bit to accommodate special handling
    of the “help” option even when normally an option wouldn’t parse
    there.

  • responsibility was offloaded from “CLI” and on-loaded to
    the “stack frame” so that the help system can manipulate a
    stack just like the main parser can.

  • blind rewrite (fourth extant?) of help screen parsing

(2 in 17.76, 17.45, 17.24)