(m4 4/J: enter "operation index" and "scope index".. (.. 3124))
Synopsis: a long time coming, an ancient stash that was waiting for a
lot of the theory and code to catch up to it (mostly near the central
algorithm of [#ze-015] exactly how we build an option parser). a lot
had to change to make this all work: arg parsing, opt parsing, operation
index, scope index.

There is one glaring problem with this one which is we have left
entirely open the problem of "recursion" (i.e "operation dependencies"),
something we want a dedicated commit for.

Wickedly and relatedly, we implement for niCLI our own "operation index"
and "scope index" rather than reuse the ones used for API. it is these
parts that need to work with recursion, and don't have those methods
yet..

Less importantly we have left open the implementation of flags of now.

• get mentions of “bespoke” out of ACS completely - “bespoken-ness”
  is in the domain of [ze] squarely, and outside of all other domains.

• [br] CLI support “arguments” outputs to different shapes now

• [br] the empty section finally gets coverage

• [br] syntax assembly gets bent for whatever happened here ..

• [ca] while we are in the neighborhood (of reading it), put a finer
  point on to what extent “future expect” is deprecated

• [ze] enter “evaluation” - get this logic out where others can use it

(2 in 16.91)