(m5 3/K: niCLI sets background image!.. (.. 3173))
Synopsis: in the interest of scope we are cutting this step just short
of *covering* the actual act of setting the background image, because
that’s going to require system stubbing which will be somewhat
involved. but through a "visual test", the niCLI does a "bare minimum"
of actually setting the background image (again)!

and we’ve covered a lot of ground: begin to appreciate that there are
“terminal adapters” and there are “image producing adapters”; both of
which leverage the [#ta-005] “magnetics” pattern but separately.

architecturally we begin to reflect this. (but this N-chotomy is still
not baked deeply into the reactive model layer..)

we experiment with new patterns that are near to (but perhaps more
complex than) [ta] task graph resoultion (in its way).

• [ac] flush pending rename: “node ticket streamer”

• [fi]<=[br] flush pending rename: docs about meta attributes

(2 in 17.06)