(m5 4/K: stub the system (.. 3175))
Theme: weaving “new ways” on top of and through the old ways:

Narrative: 

we need to stub the system so that we can test that the thing will
“probably” change the bg-image given the right circumstances, but not
actually change it when the tests are being run! (this is not the only
reason - it’s generally bad practice to reach out to external systems
like imagemagick etc when doing unit tests.)

reviewed how system stubbing worked (the older [gv] version
(now in [sy]) and the newer “inline” form that was in [gi]).
we moved the newer inline form that was in [gi] to [sy] so that
the two variants live side-by-side (in the same file even).

then, to boot we decided that we wanted yet a third variant, one
that does “diminishing pool”-based functional matching of the ARGV
passed to a `popen3` command, rather than the static matching that
the other two do. (it’s just easier & more flexible).

• [sy]<=[gi] move the “inline static” form of mocked system from
  the one to the other, so they all live together.

• [my] although we have an already existing way that we stub the
  system, we have a more direct and easy to understand way now:
  just write the system conduit..

• [sl] document workflow: checklists

(2 in 17.49)