m5 7/K: reach milestone! sets background image for real.. (.. 3192)
Synopsis: this is really just a continuation of the previous commit:
it is taxono-semantically the same. when we went for a smoke test after
the previous commit we hit some uncovered corners that we added coverage
and code for here. this lead to some refactoring that we decided for
what ever reason should get its own representation in its own commit.

  • [br] push invitation specifics back up into the
    “invocation expression” so that clients in external libs of [br]
    can reach their own hooks for customizing this “appendage”

  • [my] the “custom” node (so young) gets a thorough cleaning: the
    cached “hotstrings” thing evolved into the more general
    “common elements”, stored in the hacky “all purpose cache” thing
    in the frame.

  • [ze] niCLI invitation gets its own version of the
    “express as compound” (appropriated from [br])
    so it can be hacked over with an “invite” hook.

(2 in 18.09)