(m7 4/M: (& begin m8) flush renames (.. 3262))
Discussion: (reasons for renames are discussed in the previous commit)

we were in the middle (beginning of) of “integrating” [my] with iCLI
when we discovered that we’ve had significant changes in theory since
iCLI was started: now, “entitesque” and “primitivesque” are just a
superficial distinction. what we used to abuse “entitesque” for we now
accomplish through “formal operations”. where we used to have
“transitive operations” we now wish we didn’t have those (maybe).

so really, we need to rework iCLI to meet this change in design.
(it’s a bit daunting as it is, with the event loop etc.) so we want
it to be more intuitive and transparent.

instead of stashing the in progressed work on [my], we have created
almost a frozen snapshot of it as it was when we realized we had to
take a step back.

(2 in 17.11)