(m7 7/M: change array-based stack to LL-based.. (.. 3269))
Justification: the linked-list approach is more flexible (and also
more familiar now). with the linked list approach, a stack can always
be derived from a frame. a performer that at first needs only the frame
can later decide it needs the whole stack (at least, the stack below
the frame) without having to change construction signatures.

the converse is not true: if we represent the stack as array and pass
this array around, the particular performer is not “certain” in any
robust way that at that moment, the top of the stack has the frame
that is counterpart to the performer.

Catalyst: this comes in handy more for the next commit, where to support
complex graphs of interdependent operations, formal operations need to
be assembled with selection stacks..

(2 in 17.13)