(interlude 2/3: TONS OF RENAMES..)
Catalyst/Justification: as we were trying to unify "path tools"
(pretty path), the taxo-semantic tree of [sy] was aesthetically
upsetting:

  • flatten [sy] services - don't place find, grep, patch etc under
    filesystem -- represent them as autonomous services

    just because these services depend on the filesystem does not
    mean that we should make our hierarchy deep. those services
    maintain their own relationship to the filesystem, so we should
    ignore it.

In general that architecture came from an obsession with "scalability"
through infinitely recursive taxo-semantic trees. we have relaxed on
this obsession somewhat, now finding it reasonable to have flat lists
of files near the top of the sidesystem.
_