(interlude 3/3: unify pretty path (ish) and 2 more.. (.. 3306))
Long Narrative: the fact that iCLI zerk wasn’t using a “proper”
expression agent finally came back to bite us: a client wanted to
use the canonic `pth` function that renders a filesystem path in
a “pretty” way ([#sy-005]). the aforementioned issue (still unresolved)
drew our attention to the “non-unified” (read messy and disparate)
implementation(s) of pretty path.

“pretty path” as a family strain of functions is one of the oldest
strains in this universe and as such it has lived through multiple
“eras” of solution conventions. first there was the “impossible to
debug function-happy” solution. then there was the “just use plain
old classes” solution.

the problem was that both of these failed to address elegantly the
main problem we have with this function-family now: the fact that
its behavior is dependent on fragile, volatile characteristics of
the environment like current working directory and HOME:

  • for one thing, it should not be the case that a running tests
    changes the real PWD and thereby causes a subsequent test to break.

  • for another thing, in real life if we are doing the “pretty path”
    function on “many” paths, it’s a poor scale to have to check
    HOME and pwd for every path (on the one side); and it’s a cacheing
    problem on the other side. (this particular issue was one that the
    “first era” version tried to address, but did so in a smelly way.)

As such, we thought it was a good time to give this function its third
rewrite (this being an architectural change to the subject as well.)

BUT when we went to make changes to it we were aghast at the architecture
under [sy] filesystem. SO the changes in this commit reflect several
of the above points and more corollaries from it:

  FIRST, the bulk of this commit was tiny changes to accommodate the
  restructuring explained in the previous commit.
  (a nice by-product of this is that we have applied a regression order
  to the tests so we can more easily “fail fast” or “regress” as needed.)

  SECOND, the rewrite of “pretty path” itself went smoothly after we
  spent some serious time whiteboarding the simple arithmetic of
  calculating the benefit of transformations..

  (on the subject, “pather” is not integrated to the depth we want it
  to be integrated such that it does not use “hot” HOME and `pwd`, but
  the implementation path for this is now clear and plain.)

  THIRDLY we sort of randomly ran into this issue, which
  we parenthetically addressed:

  • design of [my] and kernel was flawed - wrote [#010] and revamped
    how caching works against testing he above happened because we
    were running the tests with quickie and not r.s and so the order
    was changed..)

  FOURTHLY, in order to discover the nasty issue that sprung from
  above described problem, we gave ‘debug!’ the ability (again)
  to let us read the message content of the events we are debugging
  in “expect event”.

(2 in 18.29, 18.59)