([dt] 6/L 2 will rewrites/removes..)
this self-supporting test that rewrites itself is wayy too much
trouble to be worth its novelty value. the current suite of tests
achieves the same end that this test did (proof that the system
works) but in the complete opposite way, through regression-friendly
tests that more or less follow the three laws, as opposed to one big
kahuna that is fragile and fails awesomely. our itch for amusing
self-referential tests is more-or-less scratched by the test we get
working in the next commit.