([ta] 02/LL explosion of FSA parsing.. (( 21 )) (.. 3510))
Synopsis: what a wild ride. we like what we came up with:
first write your grammar as a simple FSA on a whiteboard.
THEN write it as a graph viz dot file following some simple conventions.
THEN use this script to tell you what methods etc you need to add to
your parser class. (this is a summary of [#011].)

• [co] small cleanups/exposures

• [ba]<-[dt] abstract quoted string unescaping logic so it
             is reachable by [ta] as well

(2 in 16.35)