(01/R will rewrite..)
"recurse" is getting the massivest of overhauls which will be the
subject of the next perhaps 10 commits if no more.

we *might* resuscitate "manifests"; we don't know yet. we're trying
to get away from this in the spirit of convention over configuration
(we've already knocked out most of the support code for the once-were
"features" of manifests); BUT we're not sure yet whether we might
bring them back but in conjunction with what we're doing SO a
comfortable middle ground is to knock out the code but keep the
files for now..