(1/LL [ts] renames..)
About this one rename:

rename so that an executable file is out of the lib node. we used to
value keeping code under the lib over keeping the code in `bin` - now
it's the opposite: dipping into the library tree to reach an executable
file is always wrong, no matter the file. (it makes sense that a
script that does coverage-related stuff would be executable (i.e at
the very front of things).)

nowadays we don't find it *as* aggregigious for library code to
load a file that is an executable in collusion with the file
(i.e with the file's participation, knowing that it might be loaded
as a library).

regardless, coverage tools broke for us on our platform at some point,
and we probably won't get around to fixing this right now. our main
objective here was to clear out the `operations-` directory.

About other renames:

  - while we're jostling things around, we might as well give all
    of these tests regressive numbers

  - let’s not use scripts/ directory anymore

  - let’s not use the modalities/ directory any more

  - let’s not use the operations/ directory any more. [ac] obviates
    this in its way because you have to write the small adaptations
    by hand to get an ACS into a generated CLI