(06/LLL [pe] CLI cover two magnets (.. 3601))
Synopsis: we thought we could just make all the old parts snap together
like legos, but the (now) magnets had unclear interfaces because they
were not covered. and as is often the case, we assumed that the parsing
“should be” simpler than it actually needs to be, as the (now covered)
edge cases can attest to.

main work: add two test files to cover two of the three magnets.

  - [ba] state machine was written almost but not completely to accommodate
    a prototype pattern (maybe it would have just worked with `dup` but
    it wasn’t clear) - to this point we had only ever produced grammars
    inline and used them in a one-off manner.

  - [co] as documented, better ways to test emissions

  - [*]->[co] it was never appropriate to have a name like `state_` be
    part of the public API for expect event in [co]; this was a local
    idiom that got carried too far. now we have the more appropriate
    name `state_for_expect_event` that calls the locally-friendly name.

(2 in 22.51)