(20/R [sy] refactor: recording matures..)
Discussion: immediately after getting this "system recording" to work
on its first use case, we discover that it's too monolithic to
accommodate the needs of the second use case. because we don't
want to abandon the ideas we had of a simple, line-for-line, stream-
based recording, we're making the recording session be modular,
deferring to an "expression engine" for its actual work. (the next
commit after this will introduce a second expression engine that
accommodates the next use case.)

because it feels absurd to be covering this system recording utility
with tests considering how many degrees of tangent stack we are away
from doing useful work; the only confirmation here that this still
works is to run the script again from [gi] that builds the fixture.
it should look similar to the existing fixture.