(29/R     [dt] enter "identifyings".. (( 184 )) )
Design Factor, Consequence, Interesting Pattern:

The problem we were having is that “document-universal identifiers”
were perhaps a leaky abstraction to apply to non-“node of interest”
nodes. that is, when we reference nodes the than context and example
nodes to indicate that the node we are inserting should go after that
node, we cannot use the “document-universal identifier” technique; we
must use an identifying technique suited to the particular paraphernalia.

now, the “syntax definition” of the ersatz parser has a more detailed
set of choices to chose from when indicating various “identifyings”
for the component..

this “feels” much better, because now for example we don’t have to
abuse the “universal identifying string” member to for example distinguish
between a `before :each` and a `before :all`, which we knew was just
wrong to do. nor do we overload these different concerns into one
method (`mixed_universal_identifier` (or whatever)).

more at the multiple documents added here.

  - get some “test only” code out ersatz parser and
    into a test support file.

  - totally sweet overview of “pipeline” with ASCII art