(-2 [sa] 2/2 improve [br] test teardown; [*].. (( 3422 )) )
Incident report:

there is a test (or tests) that assumed (probaby reasonably)
that the string value produced by `__FILE__` for the test file
holding that code expression is a path that can be used as-is
to reference the corresponding inode on the filesystem for that
test file.

there is another test (or tests) in another sidesystem ([br])
that (before this commit) would change the `pwd` of the process for
a test, and would leave the `pwd` as-is after running the test.

when any of these particular [br] tests was executed before any of
the tests that need `__FILE__` to be "right" was *run* (not loaded
but run), the system's `pwd` was changed from what it was when these
files were loaded, effectively rendering those filesystem paths as
incorrect (for those `__FILE__` paths that report relative paths,
which seems to be the case for these paths generally because of how
we generally run tests).

`::File.expand_path` would not solve this problem but rather it
would only be a restatement of the same problem because it too
relies on the `pwd`.

using a `randomize` option at the super-sidesystem level would have
probably caught this issue sooner. (uh oh an idea)

so our fix is to follow an explict setup/teardown pattern (written
in an ad-hoc manner just for this CLI testing library in [br])

  - [br] test/cli/actions was too dependent on `::Pathname`.
    reworked it with awful indent change but meh

  - the old `Expression_Agent` name might cause headaches if
    we don't fix it universally now because of the hacky old
    way that [br] tries to peek the filesystem to load it.

  - as for .stack, oops we omitted two sidesystems in our original tally

(1 in 18.95 seconds)
.